% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc24.R
\name{Calc24}
\alias{Calc24}
\title{Game: Calculate 4 number to 24.}
\usage{
Calc24(c(a,b,c,d))
}
\arguments{
\item{a}{A integer.}

\item{b}{A integer.}

\item{c}{A integer.}

\item{d}{A integer.}
}
\value{
The final formula to calculate 24 using \code{a}, \code{b}, \code{c} and \code{d} by  +, -, *, and /.
}
\description{
Game: Calculate 4 number to 24.
}
\examples{
Calc24(c(4,4,4,4))
[1] '((4*4)+4)+4=24'
#This is the formula to calculate 24.#

Calc24(c(1,2,3,4))
[1] '((4*3)*2)*1=24'
#This is the formula to calculate 24.#
}
