% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess2.R
\name{guess2}
\alias{guess2}
\title{Game: Guess which four digits number I think!.}
\usage{
1. print guess2() 
2. Please input a four-digits number:a
}
\arguments{
\item{a}{a non-repeating four-digits for each digit from 1 to 9.}
}
\value{
If the number you input has only one digit which position and number are both correct and none of the digit has correct number and incorrect position, it will return '1A0B'.

If the number you input is correct, it will return 'Good job!!'.

When you guess too much times, it will return 'Input answer to get the answer'. Then, if you input 'answer', it will return the correct answer of this game.
}
\description{
Game: Guess which four digits number I think!.
}
\examples{
guess2()
Please input a four-digits number:4567
[1] '1A1B'
#1A means one digit we guess has correct position and number#
#1B means one digit we guess has correct number but position is incorrect.#
Please input a four-digits number:4568
'2A1B'
#We are closer to the answer.#
Please input a four-digits number:3468
'Good job!!'
#Congratulation! You win this game!#


#If we guess too much times#
Please input a four-digits number:4568
[1] 'Input answer to get the answer'
[1] '2A1B'
Please input a four-digits number:answer
[1] 3 4 6 8
#The answer is 3468.#
}
