% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texas_holdem.R
\name{texas_holdem}
\alias{texas_holdem}
\title{Game: Texas Hold'em Tool.}
\usage{
texas_holdem(c('A','B','C','D','E'))
texas_holdem(c('A','B','C','D','E','F'))
}
\arguments{
\item{A}{a card composed of a suit and a number or letter.}

\item{B}{a card composed of a suit and a number or letter.}

\item{C}{a card composed of a suit and a number or letter.}

\item{D}{a card composed of a suit and a number or letter.}

\item{E}{a card composed of a suit and a number or letter.}

\item{F}{a card composed of a suit and a number or letter.}
}
\value{
It will return the relative probability of each type of winning game.
}
\description{
Game: Texas Hold'em Tool.
}
\details{
S, H, C, and D stand for four suits: Spade, Heart, Club and Diamond.

2, 3, 4, 5, 6, 7, 8, 9, 10, J, Q, K, and A stand for the number of cards.

In this function, 'H2' stand for the two of hearts, 'HJ' stands for the jack of hearts.

The total number of cards should be 5 or 6.
}
\examples{
#If the we only know the 2 pokers we have and 3 pokers on the table, then we input the suits and number of these five pokers.#
texas_holdem(c('S4','S6','H8','H9','H7'))
                        1 pair     2 pair 3 of a kind  straight      flush full house 4 of a kind straight flush
Relative probability 0.5855689 0.08325624  0.01387604 0.3145236 0.04162812          0           0     0.01572618
#It returned the probabilitiy of each type of winning game.#


#Now, we know one more piece of card, then we add it in this function.#
texas_holdem(c('S4','S6','H8','H9','H7','SK'))
Please input a four-digits number:4567
                       1 pair 2 pair 3 of a kind straight flush full house 4 of a kind straight flush
Relative probability 0.3913043      0     0      0.173913  0          0           0              0
#It returned the new probabilitiy of each type of winning game.#
}
